requirements.txt
flask
PyYAML
pydantic
Flask-CORS
paho-mqtt